+++ plume/src/test/kotlin/io/github/plume/oss/TestDomainResources.kt
@@ -1,6 +1,5 @@
 package io.github.plume.oss
 
-import io.github.plume.oss.domain.mappers.ListMapper
 import io.github.plume.oss.drivers.IDriver
 import io.shiftleft.codepropertygraph.generated.DispatchTypes.DYNAMIC_DISPATCH
 import io.shiftleft.codepropertygraph.generated.DispatchTypes.STATIC_DISPATCH
@@ -33,8 +32,6 @@ class TestDomainResources {
             NewBlockBuilder().typeFullName(STRING_1).code(STRING_1).order(INT_1).argumentIndex(INT_1)
                 .lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1)),
             NewCallBuilder().methodFullName(STRING_1).argumentIndex(INT_1).dispatchType(DISPATCH_1)
-                .typeFullName(STRING_1)
-                .dynamicTypeHintFullName(ListMapper.stringToScalaList(STRING_1))
                 .name(STRING_1).signature(STRING_1).code(STRING_1).order(INT_1).lineNumber(Option.apply(INT_1))
                 .columnNumber(Option.apply(INT_1)),
             NewControlStructureBuilder().controlStructureType(STRING_2).code(STRING_1).lineNumber(Option.apply(INT_1))
@@ -54,7 +51,7 @@ class TestDomainResources {
             NewMetaDataBuilder().language(STRING_1).version(STRING_1).hash(Option.apply(STRING_2)),
             NewMethodParameterInBuilder().code(STRING_1).evaluationStrategy(EVAL_1).typeFullName(STRING_1)
                 .name(STRING_1).order(INT_1).lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1)),
-            NewMethodRefBuilder().methodInstFullName(Option.apply(STRING_1)).methodFullName(STRING_1).code(STRING_1)
+            NewMethodRefBuilder().methodFullName(STRING_1).code(STRING_1)
                 .order(INT_1).argumentIndex(INT_1).lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1)),
             NewMethodReturnBuilder().typeFullName(STRING_1).evaluationStrategy(EVAL_1).code(STRING_1)
                 .order(INT_1).lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1)),
@@ -69,7 +66,6 @@ class TestDomainResources {
             NewTypeDeclBuilder().name(STRING_1).fullName(STRING_1).order(INT_1).isExternal(BOOL_1),
             NewTypeParameterBuilder().name(STRING_1).order(INT_1),
             NewTypeRefBuilder().typeFullName(STRING_1)
-                .dynamicTypeHintFullName(ListMapper.stringToScalaList(STRING_1))
                 .code(STRING_1).argumentIndex(INT_1).order(INT_1).lineNumber(Option.apply(INT_1))
                 .columnNumber(Option.apply(INT_1)),
             NewTypeBuilder().name(STRING_1).fullName(STRING_1).typeDeclFullName(STRING_1),
@@ -90,8 +86,6 @@ class TestDomainResources {
                 .lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1))
         val callVertex: NewCallBuilder =
             NewCallBuilder().methodFullName(STRING_1).argumentIndex(INT_1).dispatchType(DISPATCH_1)
-                .typeFullName(STRING_1)
-                .dynamicTypeHintFullName(ListMapper.stringToScalaList(STRING_1))
                 .name(STRING_1).signature(STRING_1).code(STRING_1).order(INT_1).lineNumber(Option.apply(INT_1))
                 .columnNumber(Option.apply(INT_1))
         val localVertex: NewLocalBuilder =
@@ -131,10 +125,9 @@ class TestDomainResources {
             NewFieldIdentifierBuilder().canonicalName(STRING_1).code(STRING_1).argumentIndex(INT_1).order(INT_1)
                 .lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1))
         val methodRefVertex: NewMethodRefBuilder =
-            NewMethodRefBuilder().methodInstFullName(Option.apply(STRING_1)).methodFullName(STRING_1).code(STRING_1)
+            NewMethodRefBuilder().methodFullName(STRING_1).code(STRING_1)
                 .order(INT_1).argumentIndex(INT_1).lineNumber(Option.apply(INT_1)).columnNumber(Option.apply(INT_1))
         val typeRefVertex: NewTypeRefBuilder = NewTypeRefBuilder().typeFullName(STRING_1)
-            .dynamicTypeHintFullName(ListMapper.stringToScalaList(STRING_1))
             .code(STRING_1).argumentIndex(INT_1).order(INT_1).lineNumber(Option.apply(INT_1))
             .columnNumber(Option.apply(INT_1))
         val unknownVertex: NewUnknownBuilder =